Lib wide refactor (#117)
* refactor: lib wide refactor

* fix: tests
