+++ src/chunk/render/tilemap-square.vert
